renamed to tabor_new_driver package to feature_awg
